ionic-post-list-app
Build a responsive mobile app using Ionic that displays a list of posts using JSON Placeholder API and
should allow searching for posts. The app should load the posts from a JSON Placeholder API
endpoint and display them in a list view. Each post should display an Id, Title, and the Body. The user
should be able to search for the posts using its title